Menu renomeado para paginator e pequenos fixs

+++ src/collectors/reactionCollector.js
@@ -1,35 +1,55 @@
-const { Message, ReactionCollector: DjsReactionCollector, CollectorOptions: DjsCollectorOptions, UserResolvable } = require("discord.js");
+const { Message, ReactionCollector: DjsReactionCollector, EmojiResolvable,CollectorOptions: DjsCollectorOptions, UserResolvable, MessageEmbed } = require("discord.js");
 const {validateOptions} = require('../util/validate');
-const editMenu = async (botMessage, isBack, i, pages) => {
+const editPaginator = async (botMessage, isBack, i, pages) => {
     isBack ? (i > 0 ? --i : pages.length - 1) : (i + 1 < pages.length ? ++i : 0);
     await botMessage.edit({ embed: pages[i] });
 }
 
 module.exports = class ReactionCollector {
     /**
-     * @description This method can be used to create easier react menu, with multiples embeds pages.
-     * @param  {MenuOptions} options
+     * @description This method can be used to create easier react pagination, with multiple embeds pages.
+     * @param  {PaginatorOptions} options
+     * @param  {Message} options.botMessage - Message from Bot to create reaction collector.
+     * @param  {UserResolvable} options.user - UserResolvable who will react. 
+     * @param  {MessageEmbed[]} options.pages - Array with embeds.
+     * @param  {EmojiResolvable[]} [options.reactions] - Array with back/skip reactions.
+     * @param  {DjsCollectorOptions?} [options.collectorOptions] - Default discord.js collector options
+     * @param  {boolean?} [options.deleteReaction] - Default True - The Bot will remove reaction after user react?
+     * @param  {boolean?} [options.deleteAllReactionsWhenCollectorEnd] - Default True - The Bot will remove reaction after collector end?
      * @note {Function[]?} options.onReact cannot be set in this method. (yet)
+     * @example
+     *   const botMessage = await message.channel.send('Simple paginator...');
+     *   ReactionCollector.paginator({
+     *       botMessage,
+     *       user: message,
+     *       pages: [
+     *           new MessageEmbed({ description: 'First page content...' }),
+     *           new MessageEmbed({ description: 'Second page content...' })
+     *       ]
+     *   });
      * @returns void
      */
-    static async menu(options) {
-        const { botMessage, user, pages, collectorOptions, reactions, deleteReaction, deleteAllReactionsWhenCollectorEnd } = validateOptions(options, 'reactMenu');
+    static paginator(options) {
+        const { botMessage, user, pages, collectorOptions, reactions, deleteReaction, deleteAllReactionsWhenCollectorEnd } = validateOptions(options, 'reactPaginator');
         if (!pages || pages.length === 0)
             throw 'Invalid input: pages is null or empty';
+        if (pages.filter(page => !(page instanceof MessageEmbed)).length > 0)
+            throw 'Invalid input: pages can be only MessageEmbed array.';
 
         let i = 0;
-        await botMessage.edit({ embed: pages[0] });
-        this.question({
-            botMessage,
-            user,
-            reactions,
-            collectorOptions,
-            deleteReaction,
-            deleteAllReactionsWhenCollectorEnd,
-            onReact: [
-                async (botMessage) => await editMenu(botMessage, true, i, pages),
-                async (botMessage) => await editMenu(botMessage, false, i, pages)
-            ]
+        botMessage.edit({ embed: pages[0] }).then(() => {
+            this.question({
+                botMessage,
+                user,
+                reactions,
+                collectorOptions,
+                deleteReaction,
+                deleteAllReactionsWhenCollectorEnd,
+                onReact: [
+                    async (botMessage) => await editPaginator(botMessage, true, i, pages),
+                    async (botMessage) => await editPaginator(botMessage, false, i, pages)
+                ]
+            });
         });
     }
 
@@ -38,7 +58,7 @@ module.exports = class ReactionCollector {
      * @param  {CollectorOptions} options
      * @param  {Message} options.botMessage - Message from Bot to create reaction collector.
      * @param  {UserResolvable} options.user - UserResolvable who will react. 
-     * @param  {string[]} [options.reactions] - Array with reactions (using unicode or emoji id)
+     * @param  {EmojiResolvable[]} [options.reactions] - Array with reactions.
      * @param  {DjsCollectorOptions?} [options.collectorOptions] - Default discord.js collector options
      * @param  {Function[]?} [options.onReact] - Corresponding functions when clicking on each reaction
      * @param  {boolean?} [options.deleteReaction] - The Bot will remove reaction after user react?
@@ -65,7 +85,7 @@ module.exports = class ReactionCollector {
      * @param  {AsyncCollectorOptions} options
      * @param  {Message} options.botMessage - Message from Bot to create reaction collector.
      * @param  {UserResolvable} options.user - UserResolvable who will react. 
-     * @param  {string[]} [options.reactions] - Array with reactions (using unicode or emoji id)
+     * @param  {EmojiResolvable[]} [options.reactions] - Array with reactions.
      * @param  {DjsCollectorOptions} [options.collectorOptions] - Default discord.js collector options
      * @param  {boolean} [options.deleteReaction] - The Bot will remove reaction after user react?
      * @param  {boolean} [options.deleteAllReactionsWhenCollectorEnd] - The Bot will remove reaction after collector end?
